+++ sdk/resourcemanager/servicefabricmesh/armservicefabricmesh/ze_generated_example_secretvalue_client_test.go
@@ -100,7 +100,7 @@ func ExampleSecretValueClient_Delete() {
 }
 
 // Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/servicefabricmesh/resource-manager/Microsoft.ServiceFabricMesh/preview/2018-09-01-preview/examples/secrets/values/list.json
-func ExampleSecretValueClient_List() {
+func ExampleSecretValueClient_NewListPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -112,7 +112,7 @@ func ExampleSecretValueClient_List() {
 		log.Fatalf("failed to create client: %v", err)
 		return
 	}
-	pager := client.List("<resource-group-name>",
+	pager := client.NewListPager("<resource-group-name>",
 		"<secret-resource-name>",
 		nil)
 	for pager.More() {